Add web-test-runner generateConfig()
